+++ R/AllClasses.R
@@ -253,13 +253,14 @@ setClass("ControlEL",
 
 #' \linkS4class{ELD} class
 #'
-#' S4 class for empirical likelihood displacement.
+#' S4 class for empirical likelihood displacement. It inherits from the base
+#' type "numeric" and has a virtual slot `.Data` that contains the vector of
+#' empirical likelihood displacement values.
 #'
-#' @slot eld A numeric vector of empirical likelihood displacement values.
 #' @aliases ELD
 #' @examples
 #' showClass("ELD")
-setClass("ELD", slots = c(eld = "numeric"))
+setClass("ELD", contains = "numeric")
 
 
 #' \linkS4class{ELT} class